+++ navigator/r2-navigator-swift/EditingAction.swift
@@ -71,9 +71,9 @@ public struct EditingAction: Hashable {
 }
 
 protocol EditingActionsControllerDelegate: AnyObject {
-
     func editingActionsDidPreventCopy(_ editingActions: EditingActionsController)
-    
+    func editingActions(_ editingActions: EditingActionsController, shouldShowMenuForSelection selection: Selection) -> Bool
+    func editingActions(_ editingActions: EditingActionsController, canPerformAction action: EditingAction, for selection: Selection) -> Bool
 }
 
 
@@ -84,30 +84,50 @@ final class EditingActionsController {
 
     private let actions: [EditingAction]
     private let rights: UserRights
+    private var isEnabled = true
 
     init(actions: [EditingAction], rights: UserRights) {
         self.actions = actions
         self.rights = rights
     }
 
-    func canPerformAction(_ action: Selector) -> Bool {
-        actions.contains { $0.action == action }
+    /// Current user selection contents and frame in the publication view.
+    var selection: Selection? {
+        didSet {
+            if let selection = selection {
+                isEnabled = delegate?.editingActions(self, shouldShowMenuForSelection: selection) ?? true
+            } else {
+                isEnabled = false
+            }
+            updateSharedMenuController()
+        }
+    }
+
+    func canPerformAction(_ selector: Selector) -> Bool {
+        guard
+            isEnabled,
+            let selection = selection,
+            let action = actions.first(where: { $0.action == selector })
+        else {
+            return false
+        }
+
+        return delegate?.editingActions(self, canPerformAction: action, for: selection) ?? true
+    }
+
+    func canPerformAction(_ action: EditingAction) -> Bool {
+        canPerformAction(action.action)
     }
 
     func updateSharedMenuController() {
-        UIMenuController.shared.menuItems = actions.compactMap { $0.menuItem }
+        var items: [UIMenuItem] = []
+        if isEnabled, let selection = selection {
+            items = actions
+                .filter { delegate?.editingActions(self, canPerformAction: $0, for: selection) ?? true }
+                .compactMap { $0.menuItem }
+        }
+        UIMenuController.shared.menuItems = items
         UIMenuController.shared.update()
-    }
-    
-    
-    // MARK: - Selection
-    
-    /// Current user selection contents and frame in the publication view.
-    private var selection: (text: String, frame: CGRect)?
-    
-    /// To be called when the user selection changed.
-    func selectionDidChange(_ selection: (text: String, frame: CGRect)?) {
-        self.selection = selection
     }
 
     
@@ -115,12 +135,12 @@ final class EditingActionsController {
 
     /// Returns whether the copy interaction is at all allowed. It doesn't guarantee that the next copy action will be valid, if the license cancels it.
     var canCopy: Bool {
-        actions.contains(.copy) && rights.canCopy
+        canPerformAction(.copy) && rights.canCopy
     }
 
     /// Copies the authorized portion of the selection text into the pasteboard.
     func copy() {
-        guard let text = selection?.text else {
+        guard let text = selection?.locator.text.highlight else {
             return
         }
         guard rights.copy(text: text) else {
@@ -137,22 +157,25 @@ final class EditingActionsController {
     /// Builds a UIActivityViewController to share the authorized contents of the user selection.
     func makeShareViewController(from contentsView: UIView) -> UIActivityViewController? {
         // Peeks into the available selection contents authorized for copy.
-        guard let selection = selection else {
+        guard
+            let selection = selection,
+            let text = selection.locator.text.highlight
+        else {
             return nil
         }
-        guard canCopy, rights.canCopy(text: selection.text) else {
+        guard canCopy, rights.canCopy(text: text) else {
             delegate?.editingActionsDidPreventCopy(self)
             return nil
         }
 
-        let viewController = UIActivityViewController(activityItems: [selection.text], applicationActivities: nil)
+        let viewController = UIActivityViewController(activityItems: [text], applicationActivities: nil)
         viewController.completionWithItemsHandler = { _, completed, _, _ in
             if (completed) {
                 self.copy()
             }
         }
         viewController.popoverPresentationController?.sourceView = contentsView
-        viewController.popoverPresentationController?.sourceRect = selection.frame
+        viewController.popoverPresentationController?.sourceRect = selection.frame ?? .zero
         return viewController
     }