+++ navigator/r2-navigator-swift/VisualNavigator.swift
@@ -1,18 +1,12 @@
 //
-//  VisualNavigator.swift
-//  r2-navigator-swift
-//
-//  Created by Mickaël Menu on 06.05.19.
-//
 //  Copyright 2019 Readium Foundation. All rights reserved.
-//  Use of this source code is governed by a BSD-style license which is detailed
-//  in the LICENSE file present in the project repository where this source code is maintained.
+//  Use of this source code is governed by the BSD-style license
+//  available in the top-level LICENSE file of the project.
 //
 
 import Foundation
 import UIKit
 import R2Shared
-
 
 /// A navigator rendering the publication visually on-screen.
 public protocol VisualNavigator: Navigator {